+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 
 ### Bug fixes
 
+* [#2807](https://github.com/bbatsov/rubocop/pull/2807): `Style/WordArray` checks for commas and quotes on percent style arrays ([@karlhungus][])
 * Fix auto-correction of `not` with parentheses in `Style/Not`. ([@lumeet][])
 * [#2784](https://github.com/bbatsov/rubocop/issues/2784): RuboCop can inspect `super { ... }` and `super(arg) { ... }`. ([@alexdowad][])
 * [#2781](https://github.com/bbatsov/rubocop/issues/2781): `Performance/RedundantMerge` doesn't flag calls to `#update`, since many classes have methods by this name (not only `Hash`). ([@alexdowad][])